move test to own package
