+++ xds/internal/xdsclient/xdsclient_test.go
@@ -19,16 +19,9 @@
 package xdsclient_test
 
 import (
-	"encoding/json"
-	"fmt"
 	"testing"
 
-	"github.com/google/uuid"
-	"google.golang.org/grpc/internal/envconfig"
 	"google.golang.org/grpc/internal/grpctest"
-	"google.golang.org/grpc/internal/testutils/stats"
-	"google.golang.org/grpc/internal/xds/bootstrap"
-	"google.golang.org/grpc/xds/internal/xdsclient"
 )
 
 type s struct {
@@ -37,49 +30,4 @@ type s struct {
 
 func Test(t *testing.T) {
 	grpctest.RunSubTests(t, s{})
-}
-
-var origBootstrapContent string
-
-func init() {
-	bs, err := bootstrap.NewContentsForTesting(bootstrap.ConfigOptionsForTesting{
-		Servers: []byte(fmt.Sprintf(`[{
-			"server_uri": %q,
-			"channel_creds": [{"type": "insecure"}]
-		}]`, "non-existent-management-server")),
-		Node: []byte(fmt.Sprintf(`{"id": "%s"}`, uuid.New().String())),
-		CertificateProviders: map[string]json.RawMessage{
-			"cert-provider-instance": json.RawMessage("{}"),
-		},
-	})
-	if err != nil {
-		panic(fmt.Sprintf("Failed to create bootstrap configuration: %v", err))
-	}
-
-	origBootstrapContent = envconfig.XDSBootstrapFileContent
-	envconfig.XDSBootstrapFileContent = string(bs)
-}
-
-// TestDefaultPool_LazyLoadBootstrapConfig verifies that the DefaultPool
-// lazily loads the bootstrap configuration from environment variables when
-// an xDS client is created for the first time.
-func (s) TestDefaultPool_LazyLoadBootstrapConfig(t *testing.T) {
-	defer func() {
-		envconfig.XDSBootstrapFileContent = origBootstrapContent
-		xdsclient.DefaultPool.UnsetBootstrapConfigForTesting()
-	}()
-
-	if cfg := xdsclient.DefaultPool.BootstrapConfigForTesting(); cfg != nil {
-		t.Fatalf("DefaultPool.BootstrapConfigForTesting() = %v, want nil", cfg)
-	}
-
-	_, closeFunc, err := xdsclient.DefaultPool.NewClient(t.Name(), &stats.NoopMetricsRecorder{})
-	if err != nil {
-		t.Fatalf("Failed to create xDS client: %v", err)
-	}
-	defer closeFunc()
-
-	if xdsclient.DefaultPool.BootstrapConfigForTesting() == nil {
-		t.Fatalf("DefaultPool.BootstrapConfigForTesting() = nil, want non-nil")
-	}
 }
